Revert "Finlaymath"
