+++ .gitignore
@@ -20,4 +20,5 @@ out/
 *.log
 imagedump/
 field images/
+
 null*
