--- Vision2019-master/visionParams.properties
@@ -1,17 +1,0 @@
-#
-#Thu Nov 15 22:30:00 EST 2018
-imgDumpPath=imagedump/TrackerboxReloadedImages/
-maxSaturation=255
-minHue=0
-erodeDilateIterations=5
-minArea=0.01
-maxHue=116
-resolution=640x480
-aspectRatioThresh=0.3
-CameraSelect=-1
-maxValue=255
-imgDumpWait=1.0
-minValue=164
-imageFile=/home/thatmadhacker/programming/2706/TrackerboxReloaded/imagedump/unzip/04766_raw.png
-minSaturation=113
-distToCentreImportance=0.5
