+++ build.gradle
@@ -1,11 +1,12 @@
 // Do not mess with the following 6 lines
 plugins {
-    id 'java'
-    id 'application'
-    id 'com.github.johnrengelman.shadow' version '4.0.3' apply false
+  id 'com.github.johnrengelman.shadow' version '1.2.4'
 }
 
-apply plugin: 'com.github.johnrengelman.shadow'
+
+apply plugin: 'java'
+apply plugin: 'application'
+apply plugin: "jacoco"
 
 // Change the line below if you change the name of your main Java class
 // Comment this line and uncomment the line after it to build/run the Server, and vice versa to run the main vision program
@@ -46,17 +47,19 @@ task settestType{
 // Shouldn't need to change anything below this point
 apply from: 'dependencies.gradle'
 
+
+
 sourceSets {
     main {
         java {
             srcDir 'src/main/java/'
         }
     }
-//    test {
-//        java{
-//	    srcDir 'src/test/java/'
-//        }
-//    }
+    test {
+        java{
+	    srcDir 'src/test/java/'
+        }
+    }
 }
 
 repositories {
@@ -64,29 +67,15 @@ repositories {
     maven {
       url "http://first.wpi.edu/FRC/roborio/maven/release"
     }
-    flatDir {
-        dirs '/usr/local/frc/java', '.'
-    }
 }
 
 dependencies {
-  //compile ntcoreDep()
-  //compile 'org.opencv:opencv-java:3.1.0'
-  compile name: 'opencv-344'
-  compile name: 'ntcore'
-  compile name: 'cameraserver'
-  compile name: 'cscore'
-  compile 'edu.wpi.first.cameraserver:cameraserver-java:+:'
+  compile ntcoreDep()
+  compile 'org.opencv:opencv-java:3.1.0'
   compile 'junit:junit:4.12'
-  compile name: 'wpiutil'
-  compile name: 'wpilibj'
-  //compile cscoreDep()
-  //testCompile 'junit:junit:4.12'
-  //testCompile tntcoreDep()
-  //testCompile 'org.opencv:opencv-java:3.1.0'
-  //testCompile name: 'opencv-344'
-  //testCompile 'edu.wpi.first.cameraserver:cameraserver-java:+:'
-  //testCompile tcscoreDep()
+  testCompile 'junit:junit:4.12'
+  testCompile tntcoreDep()
+  testCompile 'org.opencv:opencv-java:3.1.0'
 }
 
 jar {
@@ -161,21 +150,30 @@ task zipOutput(type: Zip) {
 
 
 
-//tasks.withType(Test) {
-//    systemProperty "java.library.path", "${buildDir}/tests/opencv"
-//    scanForTestClasses = false
-//    include "**/*Test.class"
-//}
+tasks.withType(Test) {
+    systemProperty "java.library.path", "${buildDir}/tests/opencv"
+    scanForTestClasses = false
+    include "**/*Test.class"
+}
+
+
+
+jacoco {
+    toolVersion = "0.8.1"
+    reportsDir = file("$buildDir/JacocoReportDir")
+}
+
+
 
 build.dependsOn setbuildType
-//test.dependsOn settestType
+test.dependsOn settestType
 
 distZip.dependsOn unzipOpenCv
 distTar.dependsOn unzipOpenCv
 
 applicationDefaultJvmArgs = ["-Djava.library.path=${openCvUnzipLocation}"]
 
-//test.dependsOn tunzipOpenCv
+test.dependsOn tunzipOpenCv
 
 build.dependsOn copyToOutput
 build.dependsOn zipOutput
@@ -184,7 +182,3 @@ run.dependsOn unzipOpenCv
 clean {
     delete outputDirectory
 }
-wrapper {
-    gradleVersion = '5.0'
-}
-
