--- buildScripts/clean.sh
@@ -1,2 +1,0 @@
-#!/bin/bash
-./gradlew clean -PbuildType=arm-raspbian -PtestType=arm-raspbian
