+++ buildScripts/raspbian-linuxtest.sh
@@ -1,2 +1,2 @@
 #!/bin/bash
-./gradlew build -PbuildType=arm-raspbian -PtestType=linux-x86_64
+./gradlew build -PbuildType=arm-raspian -PtestType=linux-x86_64
