--- master.cf
@@ -1,8 +1,0 @@
-config:
-  allowOverride=true
-  
-  
-enabled:
-  tape=true
-  cargo=false
-  hatch=false
