+++ pi_scripts/installVisionProcessOnPi.sh
@@ -25,5 +25,11 @@ fi
 # TODO Explain-y comments
 
 scp ${PI_USER}@${PI_ADDR} pi_scripts/vision.service ${PI_USER}@${PI_ADDR}:/tmp
-ssh ${PI_USER}@${PI_ADDR} "sudo cp /tmp/vision.service /etc/systemd/system && sudo systemctl stop vision.service && sudo systemctl disable vision.service && sudo systemctl daemon-reload && sudo systemctl enable vision.service && sudo systemctl start vision.service"
+ssh ${PI_USER}@${PI_ADDR} "sudo cp /tmp/vision.service /etc/systemd/system"
+ssh ${PI_USER}@${PI_ADDR} "sudo systemctl stop vision.service"
+ssh ${PI_USER}@${PI_ADDR} "sudo systemctl disable vision.service"
+ssh ${PI_USER}@${PI_ADDR} "sudo systemctl daemon-reload"
+ssh ${PI_USER}@${PI_ADDR} "sudo systemctl enable vision.service"
+ssh ${PI_USER}@${PI_ADDR} "sudo systemctl start vision.service"
+
