+++ src/main/java/ca/team2706/vision/vision2019/ImageDumpScheduler.java
@@ -23,9 +23,9 @@ public class ImageDumpScheduler implements Runnable{
 				}
 				
 				try {
-					Main.imgDump(b.getRaw(), "raw",b.getTimeStamp(),b.getParams());
-					Main.imgDump(b.getBinMask(), "binMask", b.getTimeStamp(),b.getParams());
-					Main.imgDump(b.getOutput(), "output", b.getTimeStamp(),b.getParams());
+					Main.imgDump(b.getRaw(), "raw",b.getTimeStamp());
+					Main.imgDump(b.getBinMask(), "binMask", b.getTimeStamp());
+					Main.imgDump(b.getOutput(), "output", b.getTimeStamp());
 				} catch (Exception e) {
 					//Non fatal error
 				}
