+++ src/main/java/ca/team2706/vision/vision2019/Main.java
@@ -1,67 +1,82 @@
 package ca.team2706.vision.vision2019;
 
+import java.awt.Color;
+import java.awt.Graphics;
 import java.awt.image.BufferedImage;
 import java.awt.image.DataBufferByte;
 import java.io.ByteArrayInputStream;
 import java.io.File;
+import java.io.FileInputStream;
+import java.io.FileOutputStream;
 import java.io.IOException;
 import java.io.PrintWriter;
+import java.nio.file.Files;
+import java.nio.file.Paths;
+import java.nio.file.StandardCopyOption;
 import java.util.ArrayList;
-import java.util.List;
-import java.util.HashMap;
-import java.util.Map;
+import java.util.Properties;
+import java.util.Scanner;
 
 import javax.imageio.ImageIO;
 
 import org.opencv.core.CvType;
 import org.opencv.core.Mat;
 import org.opencv.core.MatOfByte;
-import org.opencv.core.MatOfPoint;
 import org.opencv.core.Rect;
 import org.opencv.core.Size;
 import org.opencv.imgcodecs.Imgcodecs;
+import org.opencv.imgproc.Imgproc;
+import org.opencv.videoio.VideoCapture;
+import org.opencv.videoio.VideoWriter;
+import org.opencv.videoio.Videoio;
 
 import edu.wpi.first.wpilibj.networktables.NetworkTable;
 
 public class Main {
-
+	
+	public static String filename = "";
 	public static ParamsSelector selector;
+	public static int timestamp = 0;
 	public static File timestampfile;
+	public static BufferedImage currentImage;
+	public static VideoCapture camera;
+	public static VisionData lastData;
+	public static boolean process = true;
+	public static boolean showMiddle = false;
+	public static boolean useCamera = true;
+	public static Mat frame;
 
-	public static List<MainThread> threads = new ArrayList<MainThread>();
-
-	public static void reloadConfig() {
-		visionParamsList.clear();
-		loadVisionParams();
-		for (MainThread thread : threads) {
-			String name = thread.visionParams.name;
-			boolean found = false;
-			for (VisionParams params : visionParamsList) {
-				if (!found) {
-					if (params.name.equals(name)) {
-						thread.updateParams(params);
-						found = true;
-					}
-				}
-
-			}
-
-		}
+	public static void setFrame(Mat f) {
+		frame = f;
 	}
 
 	// Camera Type (set in visionParams.properties)
 	// Set to 1 for USB camera, set to 0 for webcam, I think 0 is USB if
 	// there is no webcam :/
 	/** The vision parameters, this is used by the vision pipeline **/
-	public static List<VisionParams> visionParamsList = new ArrayList<VisionParams>();
+	public static VisionParams visionParams = new VisionParams();
+	/**
+	 * The vision NetworkTable, this is used to publish vision data to the
+	 * RoboRIO
+	 **/
+	public static NetworkTable visionTable;
+	/**
+	 * This is the number of seconds between dumping images to a usb stick, this
+	 * is taken from the vision parameters
+	 **/
+	public static double seconds_between_img_dumps;
+	/**
+	 * This is the current number of seconds, this resets every time it is over
+	 * the seconds_between_img_dumps
+	 **/
+	public static double current_time_seconds;
+	/** The directory that images are dumped to **/
+	public static String outputPath;
 
 	/**
 	 * A class to hold calibration parameters for the image processing algorithm
 	 */
 	public static class VisionParams {
-
-		int group;
-
 		/** This is the minimum hue that the pipeline will recognize **/
 		int minHue;
 		/** This is the maximum hue that the pipeline will recognize **/
@@ -75,22 +90,22 @@ public class Main {
 		/** This is the maximum value that the pipeline will recognize **/
 		int maxValue;
 		/**
-		 * This is how many times the pipeline will erode dilate the camera image
+		 * This is how many times the pipeline will erode dilate the camera
+		 * image
 		 **/
 		int erodeDilateIterations;
-
-		double slope, yIntercept;
-
+		/** This is the id of the camera that will be used to get images **/
+		int cameraSelect;
 		/**
-		 * The threshold to detect one large cube as 2 cubes, this is a value between 0
-		 * and 1
+		 * The threshold to detect one large cube as 2 cubes, this is a value
+		 * between 0 and 1
 		 **/
 		double aspectRatioThresh;
 		/** The minimum area that a target can have and still be recognized **/
 		double minArea;
 		/**
-		 * How important it is for a target to be close to the center of the image, this
-		 * will change depending on how well we can turn
+		 * How important it is for a target to be close to the center of the
+		 * image, this will change depending on how well we can turn
 		 **/
 		double distToCentreImportance;
 		/** The width to resize the image from the camera to **/
@@ -98,25 +113,12 @@ public class Main {
 		/** The height to resize the image from the camera to **/
 		int height;
 		/**
-		 * The size to resize the image from the camera to, this is just the width and
-		 * the height values
+		 * The size to resize the image from the camera to, this is just the
+		 * width and the height values
 		 **/
 		Size sz;
 		/** This is the image to be processed if the selected camera is -1 **/
 		String imageFile;
-
-		public NetworkTable table;
-
-		public String name;
-
-		public String outputPath;
-
-		public double secondsBetweenImageDumps;
-		
-		public boolean enabled;
-		
-		public String type,identifier;
-		
 	}
 
 	/**
@@ -126,11 +128,6 @@ public class Main {
 	public static class VisionData {
 
 		public static class Target {
-
-			double distance;
-
-			MatOfPoint contour;
-
 			/** The x center of the target in the image **/
 			int xCentre;
 			/**
@@ -144,8 +141,8 @@ public class Main {
 			 **/
 			double yCentreNorm;
 			/**
-			 * A value between 0 and 1 representing the percentage of the image the target
-			 * takes up
+			 * A value between 0 and 1 representing the percentage of the image
+			 * the target takes up
 			 **/
 			double areaNorm; // [0,1] representing how much of the screen it
 								// occupies
@@ -154,10 +151,10 @@ public class Main {
 		}
 
 		/** The List of all the targets in the image **/
-		List<Target> targetsFound = new ArrayList<Target>();
+		ArrayList<Target> targetsFound = new ArrayList<Target>();
 		/**
-		 * The target that is the most appealing, how it is chosen depends on the
-		 * distToCenterImportance value in the vision parameters
+		 * The target that is the most appealing, how it is chosen depends on
+		 * the distToCenterImportance value in the vision parameters
 		 **/
 		Target preferredTarget;
 		/** The image that contains the targets **/
@@ -165,46 +162,27 @@ public class Main {
 		/** The frames per second **/
 		public Mat binMask = new Mat();
 		public double fps;
-		public VisionParams params;
 	}
 
 	/**
 	 * Initilizes the Network Tables WARNING! Change 127.0.0.1 to the robot ip
 	 * before it is on master or it will not be fun :)
 	 */
-	public static void initNetworkTables(String ip) {
+	public static void initNetworkTables() {
 		// Tells the NetworkTable class that this is a client
 		NetworkTable.setClientMode();
 		// Sets the interval for updating NetworkTables
 		NetworkTable.setUpdateRate(0.02);
-
-		boolean use_GUI = true;
-		
-		// If on Linux don't use guis
-		if (System.getProperty("os.arch").toLowerCase().indexOf("arm") != -1) {
-			use_GUI = false;
-		}
-		
-		if (!use_GUI && ip.equals("")) {
-
-			// Sets the team number
-			NetworkTable.setTeam(2706); // Use this for the robit
-			// Enables DSClient
-			NetworkTable.setDSClientEnabled(true); // and this for the robit
-
-		} else {
-
-			if(ip.equals("")) {
-				ip = "localhost";
-			}
-			
-			// Sets the IP adress to connect to
-			NetworkTable.setIPAddress(ip); // Use this for testing
-
-		}
-
+		// Sets the team number
+		//NetworkTable.setTeam(2706); // Use this for the robit
+		// Enables DSClient
+		NetworkTable.setDSClientEnabled(true); // and this for the robit
+		// Sets the IP adress to connect to
+		NetworkTable.setIPAddress("localhost"); //Use this for testing
 		// Initilizes NetworkTables
 		NetworkTable.initialize();
+		// Sets the vision table to the "vision" table that is in NetworkTables
+		visionTable = NetworkTable.getTable("vision");
 	}
 
 	/**
@@ -212,102 +190,99 @@ public class Main {
 	 **/
 
 	public static void loadVisionParams() {
+		// Initilizes the properties
+		Properties properties = new Properties();
 		try {
+			// The file input stream that contains the vision parameters data
+			// from visionParams.properties
+			FileInputStream in = new FileInputStream("visionParams.properties");
+			// Tells the properties to use the file input stream
+			properties.load(in);
+			// Sets the selected camera to the selected camera in the properties
+			// file
+			visionParams.cameraSelect = Integer.valueOf(properties.getProperty("CameraSelect"));
+			// Sets the minimum hue to the minimum hue in the properties file
+			visionParams.minHue = Integer.valueOf(properties.getProperty("minHue"));
+			// Sets the maximum hue to the maximum hue in the properties file
+			visionParams.maxHue = Integer.valueOf(properties.getProperty("maxHue"));
+			// Sets the minimum saturation to the minimum saturation in the
+			// properties file
+			visionParams.minSaturation = Integer.valueOf(properties.getProperty("minSaturation"));
+			// Sets the maximum saturation to the maximum saturation in the
+			// properties file
+			visionParams.maxSaturation = Integer.valueOf(properties.getProperty("maxSaturation"));
+			// Sets the minimum value to the minimum value in the properties
+			// file
+			visionParams.minValue = Integer.valueOf(properties.getProperty("minValue"));
+			// Sets the maximum value to the maximum value in the properties
+			// file
+			visionParams.maxValue = Integer.valueOf(properties.getProperty("maxValue"));
+			// Sets the minimum area to the minimum area in the properties file
+			visionParams.minArea = Double.valueOf(properties.getProperty("minArea"));
+			// Sets the erode dilate iterations to the erode dilation iterations
+			// in the properties file
+			visionParams.erodeDilateIterations = Integer.valueOf(properties.getProperty("erodeDilateIterations"));
 
-			File configFile = new File("/boot/visionParams.properties");
-
-			List<String> lists = ConfigParser.listLists(configFile);
-
-			for (String s : lists) {
-
-				VisionParams visionParams = new VisionParams();
-
-				Map<String, String> data = ConfigParser.getProperties(configFile, s);
-
-				visionParams.name = s;
-
-				visionParams.minHue = Integer.valueOf(data.get("minHue"));
-				visionParams.maxHue = Integer.valueOf(data.get("maxHue"));
-				visionParams.minSaturation = Integer.valueOf(data.get("minSaturation"));
-				visionParams.maxSaturation = Integer.valueOf(data.get("maxSaturation"));
-				visionParams.minValue = Integer.valueOf(data.get("minValue"));
-				visionParams.maxValue = Integer.valueOf(data.get("maxValue"));
-
-				visionParams.aspectRatioThresh = Double.valueOf(data.get("aspectRatioThresh"));
-
-				visionParams.distToCentreImportance = Double.valueOf(data.get("distToCenterImportance"));
-
-				visionParams.imageFile = data.get("imageFile");
-
-				visionParams.minArea = Double.valueOf(data.get("minArea"));
-
-				visionParams.erodeDilateIterations = Integer.valueOf(data.get("erodeDilateIterations"));
-
-				String resolution = data.get("resolution");
-
-				visionParams.width = Integer.valueOf(resolution.split("x")[0]);
-				visionParams.height = Integer.valueOf(resolution.split("x")[1]);
-
-				// Set the vision parameters size
-				visionParams.sz = new Size(visionParams.width, visionParams.height);
-
-				visionParams.table = NetworkTable.getTable("vision-" + s);
-
-				visionParams.outputPath = data.get("imgDumpPath");
-
-				visionParams.secondsBetweenImageDumps = Double.valueOf(data.get("imgDumpTime"));
-
-				visionParams.slope = Double.valueOf(data.get("slope"));
-
-				visionParams.yIntercept = Double.valueOf(data.get("yIntercept"));
-
-				visionParams.group = Integer.valueOf(data.get("group"));
-				
-				visionParams.type = data.get("type");
-				
-				visionParams.identifier = data.get("identifier");
-				
-				visionParamsList.add(visionParams);
-
+			// Sets the aspect ratio threshold to the aspect ratio threshold in
+			// the properties file
+			visionParams.aspectRatioThresh = Double.valueOf(properties.getProperty("aspectRatioThresh"));
+			// Sets the distance to center importance to the distance to center
+			// importance in the properties file
+			visionParams.distToCentreImportance = Double.valueOf(properties.getProperty("distToCentreImportance"));
+			// Sets the output path for dumping images to the image dump path in
+			// the properties file
+			outputPath = properties.getProperty("imgDumpPath");
+			// Sets the seconds between image dumps to the seconds between image
+			// dumps in the properties file
+			seconds_between_img_dumps = Double.valueOf(properties.getProperty("imgDumpWait"));
+			// Sets the image file path to the image file path in the properties
+			// file, this is only used if the selected camera is -1
+			visionParams.imageFile = properties.getProperty("imageFile");
+			if (outputPath.endsWith("/") || outputPath.endsWith("\\")) {
+				timestampfile = new File(outputPath + "time.stamp");
+			} else {
+				timestampfile = new File(outputPath + "/" + "time.stamp");
 			}
-			
-			sendVisionParams();
+			// if the file, take the timestamp from there
+			if (timestampfile.exists()) {
+				Scanner s = new Scanner(timestampfile);
+				try{
+					timestamp = Integer.valueOf(s.nextLine()).intValue();
+				}catch(Exception e){
+					e.printStackTrace();
+				}
+				timestamp++;
+				s.close();
+			}
 
+			// Gets the resolution from the properties file
+			String resolution = properties.getProperty("resolution");
+			if (resolution.equals("320x240")) {
+				// Sets the resolution to 320 by 240
+				visionParams.width = 320;
+				visionParams.height = 240;
+			} else if (resolution.equals("640x480")) {
+				// Sets the resolution to 640 by 480
+				visionParams.width = 640;
+				visionParams.height = 480;
+			} else if (resolution.equals("160x120")) {
+				// Sets the resolution to 160 by 120
+				visionParams.width = 160;
+				visionParams.height = 120;
+			} else if (resolution.equals("80x60")) {
+				// Sets the resolution to 80 by 60
+				visionParams.width = 80;
+				visionParams.height = 60;
+			} else {
+				// Throws a error that the the resolution is invalid
+				throw new IllegalArgumentException("Error: " + properties.getProperty("resolution")
+						+ " is not a supported resolution.\n" + "Allowed: 80x60, 160x120, 320x240, 640x480.");
+			}
 		} catch (Exception e1) {
 			e1.printStackTrace();
 			System.err.println("\n\nError reading the params file, check if the file is corrupt?");
 			System.exit(1);
 		}
-	}
-	
-	private static void sendVisionParams() {
-		
-		for(VisionParams params : visionParamsList) {
-			
-			NetworkTable visionTable = NetworkTable.getTable("vision-" + params.name+"/params");
-			
-			visionTable.putNumber("group", params.group);
-			visionTable.putNumber("yIntercept", params.yIntercept);
-			visionTable.putNumber("slope", params.slope);
-			visionTable.putNumber("secondsBetweenImageDumps", params.secondsBetweenImageDumps);
-			visionTable.putNumber("height", params.height);
-			visionTable.putNumber("width", params.width);
-			visionTable.putNumber("erodeDilateIterations", params.erodeDilateIterations);
-			visionTable.putNumber("minArea", params.minArea);
-			visionTable.putString("imageFile", params.imageFile);
-			visionTable.putNumber("distToCenterImportance", params.distToCentreImportance);
-			visionTable.putNumber("aspectRatioThresh", params.aspectRatioThresh);
-			visionTable.putString("type", params.type);
-			visionTable.putString("identifier", params.identifier);
-			visionTable.putNumber("minHue", params.minHue);
-			visionTable.putNumber("maxHue", params.maxHue);
-			visionTable.putNumber("minSaturation", params.minSaturation);
-			visionTable.putNumber("maxSaturation", params.maxSaturation);
-			visionTable.putNumber("minValue", params.minValue);
-			visionTable.putNumber("maxValue", params.maxValue);
-			
-		}
-		
 	}
 
 	/**
@@ -315,67 +290,69 @@ public class Main {
 	 * 
 	 **/
 	public static void saveVisionParams() {
-
+		// Initilizes the properties object
+		Properties properties = new Properties();
 		try {
-
-			for (VisionParams params : visionParamsList) {
-
-				saveVisionParams(params);
-
-			}
-
+			// Sets the camera select property in the file to the camera select
+			// value
+			properties.setProperty("CameraSelect", String.valueOf(visionParams.cameraSelect));
+			// Sets the minimum hue property in the file to the minimum hue
+			// value
+			properties.setProperty("minHue", String.valueOf(visionParams.minHue));
+			// Sets the maximum hue property in the file to the maximum hue
+			// value
+			properties.setProperty("maxHue", String.valueOf(visionParams.maxHue));
+			// Sets the minimum saturation property in the file to the minimum
+			// saturation value
+			properties.setProperty("minSaturation", String.valueOf(visionParams.minSaturation));
+			// Sets the maximum saturation property in the file to the maximum
+			// saturation value
+			properties.setProperty("maxSaturation", String.valueOf(visionParams.maxSaturation));
+			// Sets the minimum value property in the file to the minimum value
+			// value
+			properties.setProperty("minValue", String.valueOf(visionParams.minValue));
+			// Sets the maximum value property in the file to the maximum value
+			// value
+			properties.setProperty("maxValue", String.valueOf(visionParams.maxValue));
+			// Sets the erode dilate iterations property in the file to the
+			// erode dilate iterations value
+			properties.setProperty("erodeDilateIterations", String.valueOf(visionParams.erodeDilateIterations));
+			// Sets the minimum area property in the file to the minimum area
+			// value
+			properties.setProperty("minArea", String.valueOf(visionParams.minArea));
+			// Sets the aspect ratio threshold property in the file to the
+			// aspect ratio threshold value
+			properties.setProperty("aspectRatioThresh", String.valueOf(visionParams.aspectRatioThresh));
+			// Sets the distance to center importance property in the file to
+			// the distance to center importance value
+			properties.setProperty("distToCentreImportance", String.valueOf(visionParams.distToCentreImportance));
+			// Sets the image file property in the file to the image file value
+			properties.setProperty("imageFile", visionParams.imageFile);
+			// Sets the resolution property in the file to the resolution value
+			properties.setProperty("resolution", visionParams.width + "x" + visionParams.height);
+			// Sets the image dumping interval property in the file to the image
+			// dumping interval value
+			properties.setProperty("imgDumpWait", String.valueOf(seconds_between_img_dumps));
+			// Sets the image dumping path property in the file to the image
+			// dumping path value
+			properties.setProperty("imgDumpPath", outputPath);
+			// Initilizes the output stream to the vision parameters file
+			FileOutputStream out = new FileOutputStream("visionParams.properties");
+			// Dumps the properties to the output stream
+			properties.store(out, "");
 		} catch (Exception e1) {
 			e1.printStackTrace();
 			System.exit(1);
 		}
 	}
 
-	public static void saveVisionParams(VisionParams params) throws Exception {
-		Map<String, String> data = new HashMap<String, String>();
-
-		data.put("type", params.type);
-		data.put("identifier", params.identifier);
-		data.put("minHue", String.valueOf(params.minHue));
-		data.put("maxHue", String.valueOf(params.maxHue));
-		data.put("minSaturation", String.valueOf(params.minSaturation));
-		data.put("maxSaturation", String.valueOf(params.maxSaturation));
-		data.put("minValue", String.valueOf(params.minValue));
-		data.put("maxValue", String.valueOf(params.maxValue));
-
-		data.put("aspectRatioThresh", String.valueOf(params.aspectRatioThresh));
-
-		data.put("distToCenterImportance", String.valueOf(params.distToCentreImportance));
-
-		data.put("imageFile", params.imageFile);
-
-		data.put("minArea", String.valueOf(params.minArea));
-
-		data.put("erodeDilateIterations", String.valueOf(params.erodeDilateIterations));
-
-		data.put("resolution", params.width + "x" + params.height);
-
-		data.put("imgDumpPath", params.outputPath);
-
-		data.put("imgDumpTime", String.valueOf(params.secondsBetweenImageDumps));
-
-		data.put("slope", String.valueOf(params.slope));
-
-		data.put("yIntercept", String.valueOf(params.yIntercept));
-
-		data.put("group", String.valueOf(params.group));
-
-		ConfigParser.saveList(new File("visionParams.properties"), params.name, data);
-	}
-
 	/**
-	 * Turns all the vision data into packets that kno da wae to get to the robo rio
-	 * :]
+	 * Turns all the vision data into packets that kno da wae to get to the robo
+	 * rio :]
 	 *
 	 * @param visionData
 	 */
 	public static void sendVisionDataOverNetworkTables(VisionData visionData) {
-
-		NetworkTable visionTable = visionData.params.table;
 
 		// Sends the data
 		// Puts the fps into the vision table
@@ -389,15 +366,14 @@ public class Main {
 			visionTable.putNumber("ctrX", visionData.preferredTarget.xCentreNorm);
 			// Puts the normalized area into the vision table
 			visionTable.putNumber("area", visionData.preferredTarget.areaNorm);
-			
-			visionTable.putNumber("angle", visionData.preferredTarget.xCentreNorm*45);
 		}
 	}
 
 	/**
 	 * Converts a OpenCV Matrix to a BufferedImage :)
 	 *
-	 * @param matrix Matrix to be converted
+	 * @param matrix
+	 *            Matrix to be converted
 	 * @return Generated from the matrix
 	 * @throws IOException
 	 * @throws Exception
@@ -414,7 +390,8 @@ public class Main {
 	/**
 	 * Converts a Buffered Image to a OpenCV Matrix
 	 * 
-	 * @param Buffered Image to convert to matrix
+	 * @param Buffered
+	 *            Image to convert to matrix
 	 * @return The matrix from the buffered image
 	 */
 
@@ -427,17 +404,19 @@ public class Main {
 
 	/**
 	 * 
-	 * @param The    image to dump to a file
-	 * @param image  the image to be dumped
-	 * @param suffix the suffix to put on the file name
+	 * @param The
+	 *            image to dump to a file
+	 * @param image
+	 *            the image to be dumped
+	 * @param suffix
+	 *            the suffix to put on the file name
 	 * @throws IOException
 	 */
 
-	public static void imgDump(BufferedImage image, String suffix, int timestamp, VisionParams params)
-			throws IOException {
+	public static void imgDump(BufferedImage image, String suffix, int timestamp) throws IOException {
 		// prepend the file name with the tamestamp integer, left-padded with
 		// zeros so it sorts properly
-		File output = new File(params.outputPath + String.format("%05d", timestamp) + "_" + suffix + ".png");
+		File output = new File(outputPath + String.format("%05d", timestamp) + "_" + suffix + ".png");
 		try {
 			ImageIO.write(image, "PNG", output);
 		} catch (IOException e) {
@@ -448,77 +427,251 @@ public class Main {
 		PrintWriter out = new PrintWriter(timestampfile);
 		out.println(timestamp);
 		out.close();
-
+		
 	}
 
+	public static boolean b = true;
+	
 	/**
 	 * The main method! Very important Do not delete! :] :]
 	 *
 	 * 
-	 * @param The command line arguments
+	 * @param The
+	 *            command line arguments
 	 */
-	public static void main(String[] args) throws Exception{
-
+	public static void main(String[] args) {
 		// Must be included!
 		// Loads OpenCV
-		System.loadLibrary("opencv_java344");
+		System.loadLibrary("opencv_java310");
 
-		String ip = "";
-		
-		if(args.length > 0) {
-			ip = args[0];
-		}
-		
 		// Connect NetworkTables, and get access to the publishing table
-		initNetworkTables(ip);
+		initNetworkTables();
 
 		// read the vision calibration values from file.
 		loadVisionParams();
-		
-		Map<String,String> masterConfig = ConfigParser.getProperties(new File("/boot/master.cf"), "config");
-		
-		Map<String,String> masterEnabled = ConfigParser.getProperties(new File("/boot/master.cf"), "enabled");
-		
-		String allowOverride = masterConfig.get("allowOverride");
-		
-		if(allowOverride == null || allowOverride.equals("")) {
-			
-			allowOverride = "true";
-			
+
+		try {
+			// Copys the vision parameters to a usb flash drive
+			Files.copy(Paths.get("visionParams.properties"),
+					Paths.get(outputPath + "/visionParams-" + timestamp + ".properties"),
+					StandardCopyOption.REPLACE_EXISTING);
+		} catch (IOException e2) {
+			e2.printStackTrace();
 		}
-		
-		boolean allowOverrideB = Boolean.valueOf(allowOverride);
-		
-		if(allowOverrideB)
-			NetworkTablesManager.init();
+		// Initilizes a Matrix to hold the frame
 
-		ImageDumpScheduler.start();
+		frame = new Mat();
 
-		VisionCameraServer.startServer();
+		// Open a connection to the camera
+		VideoCapture camera = null;
 
-		for (VisionParams params : visionParamsList) {
+		// Whether to use a camera, or load an image file from disk.
+		if (visionParams.cameraSelect == -1) {
+			useCamera = false;
+		}
+
+		if (useCamera) {
+			// Initilizes the camera
+			camera = new VideoCapture(visionParams.cameraSelect);
+
+			// Sets camera parameters
+			int fourcc = VideoWriter.fourcc('M', 'J', 'P', 'G');
+			camera.set(Videoio.CAP_PROP_FOURCC, fourcc);
+			camera.set(Videoio.CAP_PROP_FRAME_WIDTH, visionParams.width);
+			camera.set(Videoio.CAP_PROP_FRAME_HEIGHT, visionParams.height);
+
+			camera.read(frame);
+
+			if (!camera.isOpened()) {
+				// If the camera didn't open throw an error
+				System.err.println("Error: Can not connect to camera");
+				// Exit
+				System.exit(1);
+			}
+
+			// Set up the camera feed
+			camera.read(frame);
+		} else {
+			// load the image from file.
 			try {
-				
-				String s = masterEnabled.get(params.name);
-				
-				if(s == null || s.equals("")) {
-					s = "true";
-				}
-				
-				boolean enabled = Boolean.valueOf(s);
-				
-				params.enabled = enabled;
-				
-				VisionCameraServer.initCamera(params.type,params.identifier);
-				MainThread thread = new MainThread(params);
-				if(enabled) {
-					thread.start();
-				}
-				threads.add(thread);
-			} catch (Exception e) {
+				frame = bufferedImageToMat(ImageIO.read(new File(visionParams.imageFile)));
+			} catch (IOException e) {
 				e.printStackTrace();
+				frame = new Mat();
 			}
 		}
+		// The window to display the raw image
+		DisplayGui guiRawImg = null;
+		// The window to display the processed image
+		DisplayGui guiProcessedImg = null;
+		// Wether to open the guis
+		boolean use_GUI = true;
+		// If on Linux don't use guis
+		if (System.getProperty("os.name").toLowerCase().indexOf("raspbian") != -1) {
+			use_GUI = false;
+		}
+		// Set the vision parameters size
+		visionParams.sz = new Size(visionParams.width, visionParams.height);
+		if (use_GUI) {
+			// Resizes the frame to the vision parameters size
+			Imgproc.resize(frame, frame, visionParams.sz);
+		}
+		// Set up the GUI display windows
+		if (use_GUI) {
+			try {
+				// Initilizes the window to display the raw image
+				guiRawImg = new DisplayGui(matToBufferedImage(frame), "Raw Camera Image",true);
+				// Initilizes the window to display the processed image
+				guiProcessedImg = new DisplayGui(matToBufferedImage(frame), "Processed Image",true);
+				// Initilizes the parameters selector
+				ParamsSelector selector = new ParamsSelector(true,true);
+				guiRawImg.addKeyListener(selector);
+				guiProcessedImg.addKeyListener(selector);
+			} catch (IOException e) {
+				// means mat2BufferedImage broke
+				// non-fatal error, let the program continue
+			}
+		}
+		ImageDumpScheduler.start();
+		// Main video processing loop
+		while (b) {
+			if (useCamera) {
+				// Read the frame from the camera, if it fails try again
+				if (!camera.read(frame)) {
+					System.err.println("Error: Failed to get a frame from the camera");
+					continue;
+				}
+			} // else use the image from disk that we loaded above
+			else{
+				// load the image from file.
+	            try {
+        	        frame = bufferedImageToMat(ImageIO.read(new File(visionParams.imageFile)));
+                } catch (IOException e) {
+                	e.printStackTrace();
+                    frame = new Mat();
+                }
 
+			}
+			if (use_GUI) {
+				// Resize the frame
+				Imgproc.resize(frame, frame, visionParams.sz);
+			}
+			// Process the frame!
+			// Log when the pipeline starts
+			long pipelineStart = System.nanoTime();
+			// Process the frame
+			VisionData visionData = Pipeline.process(frame, visionParams, use_GUI);
+			// Log when the pipeline stops
+			long pipelineEnd = System.nanoTime();
+			// Selects the prefered target
+			Pipeline.selectPreferredTarget(visionData, visionParams);
+			// Creates the raw output image object
+			Mat rawOutputImg;
+			if (use_GUI) {
+				// If use gui then draw the prefered target
+				// Sets the raw image to the frame
+				rawOutputImg = frame.clone();
+				// Draws the preffered target
+				Pipeline.drawPreferredTarget(rawOutputImg, visionData);
+			} else {
+				// Sets the raw image to the frame
+				rawOutputImg = frame;
+			}
+			// Sends the data to the vision table
+			sendVisionDataOverNetworkTables(visionData);
+			lastData = visionData;
+			// display the processed frame in the GUI
+			if (use_GUI) {
+				try {
+					// May throw a NullPointerException if initializing
+					// the window failed
+					BufferedImage raw = matToBufferedImage(rawOutputImg);
+					currentImage = raw;
+					if (showMiddle) {
+						Graphics g = raw.getGraphics();
+						g.setColor(Color.RED);
+						g.fillOval(raw.getWidth() / 2 - 8, raw.getHeight() / 2 - 8, 8, 8);
+						g.dispose();
+					}
+					guiRawImg.updateImage(raw);
+					guiProcessedImg.updateImage(matToBufferedImage(visionData.binMask));
+				} catch (IOException e) {
+					// means mat2BufferedImage broke
+					// non-fatal error, let the program continue
+					continue;
+				} catch (NullPointerException e) {
+					e.printStackTrace();
+					System.out.println("Window closed");
+					Runtime.getRuntime().halt(0);
+				} catch (Exception e) {
+					// just in case
+					e.printStackTrace();
+					continue;
+				}
+			}
+			if (useCamera) {
+				// log images to file once every seconds_between_img_dumps
+				double elapsedTime = ((double) System.currentTimeMillis() / 1000) - current_time_seconds;
+				// If the elapsed time is more that the seconds between image
+				// dumps
+				// then dump images asynchronously
+				if (elapsedTime >= seconds_between_img_dumps) {
+					// Sets the current number of seconds
+					current_time_seconds = (((double) System.currentTimeMillis()) / 1000);
+					// Clones the frame
+					Mat finalFrame = frame.clone();
+					try {
+						Mat draw = finalFrame.clone();
+						Pipeline.drawPreferredTarget(draw, visionData);
+						Bundle b = new Bundle(matToBufferedImage(finalFrame), matToBufferedImage(visionData.binMask),
+								matToBufferedImage(draw), timestamp);
+						ImageDumpScheduler.schedule(b);
+						timestamp++;
+					} catch (IOException e) {
+						e.printStackTrace();
+						return;
+					}
+				}
+			}
+			// Display the frame rate onto the console
+			double pipelineTime = (((double) (pipelineEnd - pipelineStart)) / Pipeline.NANOSECONDS_PER_SECOND) * 1000;
+			System.out.printf("Vision FPS: %3.2f, pipeline took: %3.2f ms\n", visionData.fps, pipelineTime);
+		}
 	} // end main video processing loop
+
+	public static void hideMiddle() {
+		showMiddle = false;
+	}
+
+	public static void showMiddle() {
+		showMiddle = true;
+	}
+
+	public static VisionData forceProcess() {
+		Mat frame = new Mat();
+		camera.read(frame);
+		Imgproc.resize(frame, frame, visionParams.sz);
+
+		VisionData visionData = Pipeline.process(frame, visionParams, false);
+
+		Pipeline.selectPreferredTarget(visionData, visionParams);
+
+		return visionData;
+	}
+
+	public static VisionData forceProcess(Mat frame) {
+		Imgproc.resize(frame, frame, visionParams.sz);
+
+		VisionData visionData = Pipeline.process(frame, visionParams, false);
+
+		Pipeline.selectPreferredTarget(visionData, visionParams);
+
+		return visionData;
+	}
+
+	public static Mat getFrame() {
+		Mat frame = new Mat();
+		camera.read(frame);
+		return frame;
+	}
 }
