+++ src/test/java/ca/team2706/vision/vision2019/ImageDumpSchedulerTest.java
