+++ visionParams.properties
@@ -1,60 +1,17 @@
-tape:
-  minSaturation=132
-  maxValue=94
-  type=usb
-  identifier=0
-  erodeDilateIterations=0
-  distToCenterImportance=1.0
-  imgDumpPath=
-  slope=-1.3
-  resolution=640x480
-  minValue=0
-  imgDumpTime=1.0
-  maxSaturation=255
-  yIntercept=110.0
-  aspectRatioThresh=0.0
-  imageFile=/home/thatmadhacker/FRC/PhotoNASASmall/IMG_5803.JPG
-  maxHue=123
-  minHue=0
-  minArea=0.001
-  group=1
-hatch:
-  minSaturation=0
-  maxValue=255
-  type=usb
-  identifier=0
-  erodeDilateIterations=3
-  distToCenterImportance=0.5
-  imgDumpPath=
-  slope=-1.3
-  resolution=640x480
-  minValue=135
-  imgDumpTime=1.0
-  maxSaturation=116
-  yIntercept=110.0
-  aspectRatioThresh=0.0
-  imageFile=/home/thatmadhacker/FRC/PhotoNASASmall/IMG_5803.JPG
-  maxHue=57
-  minHue=0
-  minArea=0.0
-  group=0
-cargo:
-  minSaturation=22
-  maxValue=101
-  type=usb
-  identifier=0
-  erodeDilateIterations=4
-  distToCenterImportance=0.51
-  imgDumpPath=
-  slope=-1.3
-  resolution=640x480
-  minValue=69
-  imgDumpTime=1.0
-  maxSaturation=101
-  yIntercept=110.0
-  aspectRatioThresh=0.0
-  imageFile=/home/thatmadhacker/FRC/PhotoNASASmall/IMG_5803.JPG
-  maxHue=98
-  minHue=0
-  minArea=0.0
-  group=0
+#
+#Thu Nov 15 22:30:00 EST 2018
+imgDumpPath=imagedump/TrackerboxReloadedImages/
+maxSaturation=255
+minHue=0
+erodeDilateIterations=5
+minArea=0.01
+maxHue=116
+resolution=640x480
+aspectRatioThresh=0.3
+CameraSelect=-1
+maxValue=255
+imgDumpWait=1.0
+minValue=164
+imageFile=/home/thatmadhacker/programming/2706/TrackerboxReloaded/imagedump/unzip/04766_raw.png
+minSaturation=113
+distToCentreImportance=0.5
